third_party/blink/web_tests/wpt_internal/css/css-transitions/parsing/overlay-valid.html

<!DOCTYPE html>
<title>CSS Transitions: parsing overlay with valid values</title>
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/8189">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  // overlay: none | auto
  test_valid_value("overlay", "none");
  test_valid_value("overlay", "auto");
</script>
